requirements.txt
pip~=22.3.1
attrs~=23.1.0
cryptography~=40.0.2
Pillow~=9.5.0
Jinja2~=3.1.2
MarkupSafe~=2.1.3
Werkzeug~=2.3.6
click~=8.1.4
blinker~=1.6.2
itsdangerous~=2.1.2
greenlet~=2.0.2
setuptools~=65.5.0
SQLAlchemy~=2.0.18
Flask~=2.3.2
cachetools~=5.3.1
PyJWT~=2.6.0